+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 
 * nostr: add `EventBuilder::allow_self_tagging` ([Yuki Kishimoto])
 * pool: event verification cache ([Yuki Kishimoto])
+* ffi: add Mac Catalyst support in Swift package ([Yuki Kishimoto])
 
 ### Fixed